Completed All Feature in Admin

+++ app/Http/Controllers/materiController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use RealRashid\SweetAlert\Facades\Alert;
 use App\Models\Materi;
 use App\Models\Bidan;
 use App\Models\Pasien;
@@ -11,14 +12,19 @@ use Illuminate\Support\Facades\Auth;
 
 class materiController extends Controller
 {
-    public function index(){
+    public function index()
+    {
         $data['title'] = 'Data Materi';
         $data['materis'] = Materi::all();
+        $title = 'Hapus Data!';
+        $text = 'Apakah Anda Yakin?';
+        confirmDelete($title, $text);
 
         return view('admin.materi.index', compact('data'));
     }
 
-    public function indexMateri(){
+    public function indexMateri()
+    {
         $data['title'] = 'Data Materi';
         $data['materis'] = Materi::all();
 
@@ -30,7 +36,7 @@ class materiController extends Controller
             $bidan = Bidan::findOrFail($bidanId);
 
             // Pastikan bahwa $bidan->pasiens merupakan objek yang valid
-            $pasiens = $bidan-> pasiens;
+            $pasiens = $bidan->pasiens;
         }
 
         $data['pasiens'] = $pasiens;
@@ -38,40 +44,44 @@ class materiController extends Controller
         return view('bidan.materi.index', compact('data'));
     }
 
-    public function create(){
+    public function create()
+    {
         $data['title'] = 'Tambah Materi';
 
         return view('admin.materi.create', compact('data'));
     }
 
-    public function store(Request $request){
-        $request -> validate([
+    public function store(Request $request)
+    {
+        $request->validate([
             'file' => 'required|mimes:pdf|max:10240',
             'judul' => 'required',
         ]);
 
         $file = $request->file('file');
-        $fileName = time() . '.' . $file->getClientOriginalExtension();
 
-        $file->storeAs('public/uploads', $fileName);
+        $file->storeAs('public/uploads', $file->getClientOriginalName());
 
         Materi::create([
-            'file' => $fileName,
+            'file' => $file->getClientOriginalName(),
             'judul' => $request->judul,
         ]);
 
-        return redirect() -> route('materi.index');
+        Alert::success('Berhasil', 'Data Materi Berhasil Ditambahkan');
 
+        return redirect()->route('materi.index');
     }
 
-    public function edit($id){
+    public function edit($id)
+    {
         $data['title'] = 'Data Materi';
         $data['materis'] = Materi::find($id);
 
         return view('admin.materi.edit', compact('data'));
     }
 
-    public function update(Request $request, $id){
+    public function update(Request $request, $id)
+    {
         $request->validate([
             'file' => 'nullable|mimes:pdf|max:10240', // File bisa diubah atau tidak diubah
             'judul' => 'required',
@@ -81,25 +91,35 @@ class materiController extends Controller
 
         $fileToDelete = 'public/uploads/' . $materi->file;
 
-        // Alternatif menggunakan unlink
-        if (file_exists(storage_path('app/' . $fileToDelete))) {
-            unlink(storage_path('app/' . $fileToDelete));
+        $file = $request->file('file');
+        $judul = $request->judul;
+
+        if ($file) {
+            // File handling logic
+            $fileToDelete = 'public/uploads/' . $materi->file;
+
+            if (file_exists(storage_path('app/' . $fileToDelete))) {
+                unlink(storage_path('app/' . $fileToDelete));
+            }
+
+            $file->storeAs('public/uploads', $file->getClientOriginalName());
+            // Update file in the database
+            $materi->update([
+                'file' => $file->getClientOriginalName(),
+            ]);
+        } else {
+            $materi->update([
+                'judul' => $judul,
+            ]);
         }
 
-        $file = $request->file('file');
+        Alert::success('Berhasil', 'Data Materi Berhasil Diperbarui');
 
-        $file->storeAs('public/uploads', $file);
-
-        // Perbarui file dan judul di database
-        $materi->update([
-            'file' => $file,
-            'judul' => $request->judul,
-        ]);
-
-        return redirect() -> route('materi.index');
+        return redirect()->route('materi.index');
     }
 
-    public function destroy($id){
+    public function destroy($id)
+    {
         $materi = Materi::find($id);
 
         $fileToDelete = 'public/uploads/' . $materi->file;
@@ -109,8 +129,9 @@ class materiController extends Controller
             unlink(storage_path('app/' . $fileToDelete));
         }
 
-        $materi -> delete();
+        $materi->delete();
+        Alert::success('Berhasil', 'Data Materi Berhasil Dihapus');
 
-        return redirect() -> route('materi.index');
+        return redirect()->route('materi.index');
     }
 }
